refactor(core): host disambiguatedLabels next to domain labeling (#841)
Move the scale-domain label helper out of legend pixel builders into
domain-labels.ts. Scale pipeline consumers import the new home; legend
and package index re-exports keep the public path stable.
