+++ packages/core/src/pipeline/scale-style-numeric.ts
@@ -1,6 +1,6 @@
 /** Numeric style scales: size / linewidth / alpha (identity, sequential, binned, ordinal, manual). */
 import { linearTicks } from "../layout/ticks.js";
-import { disambiguatedLabels } from "../legend.js";
+import { disambiguatedLabels } from "../domain-labels.js";
 import type { StyleScale } from "../scales/style.js";
 import { encodeKey, type ScaleState } from "../scales/state.js";
 import { finiteExtent } from "../scales/train.js";
